Reorganized QuCircuit into helpers and components
