+++ frontend/src/components/quCircuit/connectionPanel.js
@@ -20,8 +20,8 @@ export default function ConnectionPanel(props) {
     const classes = useStyles();
 
     /**
-     * Apply type for autocomplete.
-     * @type {./CellLife}
+     * Apply for type checking and autocomplete.
+     * @type {CellLife}
      */
     let cellLife = props.cellLife;
