--- frontend/src/components/quCircuit/gateReference.js
@@ -1,8 +1,0 @@
-
-export default class GateReference {
-
-    constructor(index, gate) {
-        this.index = index;
-        this.gate = gate;
-    }
-}
