+++ frontend/src/helpers/quCircuit/cellData.js
