+++ frontend/src/helpers/quCircuit/cellLife.js
